Remove irrelevant part of the code-review README
One of the bullets on the code-review section is about keeping the
commits separated before merging to keep the changes and corrections
based on feedback easy to follow.

The section seems to assume that after that, a merge squash is applied
to master, but since merge squash is totally outside the ecostage
workflow at any point, the points mentioned on the bullet, though
important, are simply not applied to our case.